Move TcpClient into its own class file. Improve ROS param setting.
