+++ flir_ptu_driver/launch/ptu.launch
@@ -8,13 +8,10 @@
   <node name="robot_state_publisher" pkg="robot_state_publisher" type="state_publisher" />
 
   <node name="ptu_driver" pkg="flir_ptu_driver" type="ptu_node" ns="ptu" output="screen">
-    <!-- FLIR PTU "connection_type" is either tty or tcp. 
-    	if "tcp", it will use "ip_addr" and "tcp_port" for the connection.
-    	Otherwise, it will use the tty "port" params --> 
-    <param name="connection_type" value="tcp" /> 
-    <param name="ip_addr" type="str" value="192.168.40.195" /> 
+    <param name="connection_type" type="str" value="tcp" /> <!-- either tcp (uses ip_addr and tcp_port) or tty (uses tty port) -->
+    <param name="ip_addr" type="str" value="192.168.1.25" /> 
     <param name="tcp_port" type="int" value="4000" /> 
-    <param name="port" value="$(arg port)" /> <!-- like "/dev/ttyUSB0" -->
+    <param name="port" value="$(arg port)" /> <!-- tty port, like "/dev/ttyUSB0" -->
     <param name="test_pan_tilt_mode" type="bool" value="false"/> <!-- test mode sends random pan/tilt every few secs -->
     <remap from="state" to="/joint_states" />
   </node>
